backend/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
Flask-Bcrypt==1.0.1
gunicorn==21.2.0
Werkzeug==2.3.7
alembic==1.12.0
Flask-Mail==0.9.1
